.agents/skills/sea-transit-go-nogo/scripts/requirements.txt
pdfplumber>=0.11.0
python-dateutil>=2.9.0
